+++ source/add_extra_ac3_multichannel_audio/changelog.md
@@ -1,4 +1,9 @@
 
+**<span style="color:#56adda">0.0.6</span>**
+- add test to see if a 6 channel ac3 stream already exists with language matching that of selected TrueHD, eac3, or DTS audio - if so skip file
+- add correct stream title metadata for added ac3 audio stream
+- add fix to remove unnecessary mp4 metadata chapter title stream when output path suffix is mp4
+
 **<span style="color:#56adda">0.0.5</span>**
 - fix non-existent streams reference - change to probe_streams